refactor: move server setup to separate file, add scores
