+++ packages/backend/code/constants.ts
@@ -1,0 +1,5 @@
+// Constants for how many words of each color should be used
+export const BLUE_WORDS = 9;
+export const RED_WORDS = 8;
+export const GRAY_WORDS = 7;
+export const BLACK_WORDS = 1;
